Created starting board
Created both an 8x8 board of tiles as well as set up the board with initial 4 pieces. Setup GameController to be easy to use with themes as object prefabs can be replaced.
