+++ Othello_Unity/Assets/Prefabs/Board.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7fcbbcf763031c94fa74de2681556e67
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
