+++ Othello_Unity/Assets/Prefabs/P1_tile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e3aefac3dd1534b48b7a43cde727d493
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
