+++ Othello_Unity/Assets/Prefabs/P2_tile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 440fd525525d3c242899c154f64291f5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
