+++ Othello_Unity/Assets/Prefabs/Tile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6b03a4801cd1f0540a2f3b19b0c54742
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
