+++ Othello_Unity/Assets/Scripts.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 20d567e6210e34c419373ca3681e04c7
-NativeFormatImporter:
+guid: b4d96e16f6edc8b4bbebe47ecc0273bd
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
