+++ Othello_Unity/Assets/Sprites.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0cbdb143929672d4d9a5b838e9f34da9
+guid: f2a0d5c0253159c4085e6cb5ad9dd5a3
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
